backend/requirements.txt
flask==3.0.3
flask-cors==4.0.2
python-keycloak==4.0.0
gunicorn==23.0.0